Make children argument of the node_factory convertible to bool
As the moment, the `children` argument passed to the `node_factory`
callback of the `Trie.traverse` method is a generator.  This means
that to determine whether node has any children, the factory has to at
least attempt to generate the first child.

Change the `children` argument into an iterator with a `__bool__`
method.  It thus makes it possible to determine whether node has
children in constant time without having to iterate over them.

Fixes: https://github.com/mina86/pygtrie/pull/2